requirements.txt
flask
python-dotenv
pytesseract
Pillow
PyMuPDF
requests
